Move polar interaction to separate file / Rename argument properties to comply with standards
